requirements.txt
requests
pytest
pydantic
allure-pytest
pytest-mock
fakerplaywright==1.58.0
pytest-playwright==0.7.2
